Tighten identity-plugin contracts and remove uint8 cast warning
- Declare dtype=np.uint8 on AdaFace, ArcFace and InsightFace so the
  framework hands aligned faces in uint8 directly. Removes the per-batch
  ``batch.astype(np.uint8)`` cast and the ``RuntimeWarning: invalid value
  encountered in cast`` raised when cubic-warp overshoot pushed float32
  pixels just outside [0, 255].
- Thread the adapter's ``embedding_dim`` through the CVLFace and
  InsightFace loader factories so the empty-batch return uses the
  adapter's declared dimension instead of a hardcoded 512.
- Run the InsightFace recognition pack in a single batched
  ``recognition.get_feat(list(faces))`` call instead of one ONNX
  session.run per face — removes 16x kernel launch overhead at the
  default batch size.
- Simplify the CVLFace channel swap in ``_prepare`` to use stride
  reversal (``faces[..., ::-1]`` / ``faces[:, ::-1, :, :]``) with
  ``np.ascontiguousarray``, avoiding the fancy-index intermediate copy,
  and prefer NHWC detection so degenerate (N, 3, 3, 3) shapes are not
  silently misclassified as NCHW.
- Update the identity plugin tests to reflect the new uint8 contract on
  ``process``.

+++ plugins/extract/identity/adaface.py
@@ -31,7 +31,7 @@ class AdaFace(FacePlugin):
             input_size=adapter.input_size,
             batch_size=cfg.batch_size(),
             is_rgb=False,
-            dtype="float32",
+            dtype=np.uint8,
             scale=(0, 255),
             force_cpu=self._force_cpu,
             centering="face",
@@ -51,7 +51,7 @@ class AdaFace(FacePlugin):
 
     def process(self, batch: np.ndarray) -> np.ndarray:
         """Return AdaFace embeddings for aligned BGR face crops."""
-        return self.model.embed(batch.astype(np.uint8))
+        return self.model.embed(batch)
 
     def post_process(self, batch: np.ndarray) -> np.ndarray:
         """Ensure stored AdaFace embeddings are stable float32 unit vectors."""
